Added Upload Higscore Screen and upload function

+++ Assets/Scripts/HighScoreScripts/Score.cs
@@ -6,9 +6,9 @@ public class Score
 {
     public int id;
     public string name;
-    public float score;
+    public double score;
 
-    public Score(int id, string name, float score)
+    public Score(int id, string name, double score)
     {
         this.id = id;
         this.name = name;
